1. Has been created yandex api module. 2. Has been refactored Real Estate Config.

+++ realEstateScraperCsharp/Modules/CianOffer.cs
@@ -17,8 +17,8 @@ namespace realEstateScraperCsharp.Modules
     }
 
     class CianOfferGeoCoordinates {
-        public float? Lat { get; set; }
-        public float? Lng { get; set; }
+        public double? Lat { get; set; }
+        public double? Lng { get; set; }
     }
 
     class CianOfferGeoAddress
